Move code from experimental.go to node.go
