+++ pkg/agent/attestor/node/node.go
@@ -11,17 +11,22 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/spiffe/go-spiffe/v2/bundle/spiffebundle"
 	"github.com/spiffe/go-spiffe/v2/spiffeid"
+	agentv1 "github.com/spiffe/spire-api-sdk/proto/spire/api/server/agent/v1"
+	bundlev1 "github.com/spiffe/spire-api-sdk/proto/spire/api/server/bundle/v1"
+	"github.com/spiffe/spire-api-sdk/proto/spire/api/types"
 	"github.com/spiffe/spire/pkg/agent/catalog"
 	"github.com/spiffe/spire/pkg/agent/client"
 	"github.com/spiffe/spire/pkg/agent/plugin/keymanager"
 	"github.com/spiffe/spire/pkg/agent/plugin/nodeattestor"
 	"github.com/spiffe/spire/pkg/agent/storage"
+	"github.com/spiffe/spire/pkg/common/bundleutil"
 	"github.com/spiffe/spire/pkg/common/cryptoutil"
 	"github.com/spiffe/spire/pkg/common/idutil"
 	"github.com/spiffe/spire/pkg/common/telemetry"
 	telemetry_agent "github.com/spiffe/spire/pkg/common/telemetry/agent"
 	telemetry_common "github.com/spiffe/spire/pkg/common/telemetry/common"
 	"github.com/spiffe/spire/pkg/common/util"
+	"github.com/spiffe/spire/pkg/common/x509util"
 	"github.com/zeebo/errs"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
@@ -168,6 +173,39 @@ func (a *attestor) loadBundle() (*spiffebundle.Bundle, error) {
 	return spiffebundle.FromX509Authorities(a.c.TrustDomain, bundle), nil
 }
 
+func (a *attestor) getBundle(ctx context.Context, conn *grpc.ClientConn) (*spiffebundle.Bundle, error) {
+	updatedBundle, err := bundlev1.NewBundleClient(conn).GetBundle(ctx, &bundlev1.GetBundleRequest{})
+	if err != nil {
+		return nil, fmt.Errorf("failed to get updated bundle %w", err)
+	}
+
+	b, err := bundleutil.CommonBundleFromProto(updatedBundle)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse trust domain bundle: %w", err)
+	}
+
+	bundle, err := bundleutil.SPIFFEBundleFromProto(b)
+	if err != nil {
+		return nil, fmt.Errorf("invalid trust domain bundle: %w", err)
+	}
+
+	return bundle, err
+}
+
+func (a *attestor) getSVID(ctx context.Context, conn *grpc.ClientConn, csr []byte, attestor nodeattestor.NodeAttestor) ([]*x509.Certificate, bool, error) {
+	// make sure all of the streams are cancelled if something goes awry
+	ctx, cancel := context.WithCancel(ctx)
+	defer cancel()
+
+	stream := &ServerStream{Client: agentv1.NewAgentClient(conn), Csr: csr, Log: a.c.Log}
+
+	if err := attestor.Attest(ctx, stream); err != nil {
+		return nil, false, err
+	}
+
+	return stream.SVID, stream.Reattestable, nil
+}
+
 // Read agent SVID from data dir. If an error is encountered, it will be logged and `nil`
 // will be returned.
 func (a *attestor) readSVIDFromDisk() ([]*x509.Certificate, bool) {
@@ -266,6 +304,75 @@ func (a *attestor) serverConn(ctx context.Context, bundle *spiffebundle.Bundle) 
 	)
 }
 
+type ServerStream struct {
+	Client       agentv1.AgentClient
+	Csr          []byte
+	Log          logrus.FieldLogger
+	SVID         []*x509.Certificate
+	Reattestable bool
+	stream       agentv1.Agent_AttestAgentClient
+}
+
+func (ss *ServerStream) SendAttestationData(ctx context.Context, attestationData nodeattestor.AttestationData) ([]byte, error) {
+	return ss.sendRequest(ctx, &agentv1.AttestAgentRequest{
+		Step: &agentv1.AttestAgentRequest_Params_{
+			Params: &agentv1.AttestAgentRequest_Params{
+				Data: &types.AttestationData{
+					Type:    attestationData.Type,
+					Payload: attestationData.Payload,
+				},
+				Params: &agentv1.AgentX509SVIDParams{
+					Csr: ss.Csr,
+				},
+			},
+		},
+	})
+}
+
+func (ss *ServerStream) SendChallengeResponse(ctx context.Context, response []byte) ([]byte, error) {
+	return ss.sendRequest(ctx, &agentv1.AttestAgentRequest{
+		Step: &agentv1.AttestAgentRequest_ChallengeResponse{
+			ChallengeResponse: response,
+		},
+	})
+}
+
+func (ss *ServerStream) sendRequest(ctx context.Context, req *agentv1.AttestAgentRequest) ([]byte, error) {
+	if ss.stream == nil {
+		stream, err := ss.Client.AttestAgent(ctx)
+		if err != nil {
+			return nil, fmt.Errorf("could not open attestation stream to SPIRE server: %w", err)
+		}
+		ss.stream = stream
+	}
+
+	if err := ss.stream.Send(req); err != nil {
+		return nil, fmt.Errorf("failed to send attestation request to SPIRE server: %w", err)
+	}
+
+	resp, err := ss.stream.Recv()
+	if err != nil {
+		return nil, fmt.Errorf("failed to receive attestation response: %w", err)
+	}
+
+	if challenge := resp.GetChallenge(); challenge != nil {
+		return challenge, nil
+	}
+
+	svid, err := getSVIDFromAttestAgentResponse(resp)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse attestation response: %w", err)
+	}
+
+	if err := ss.stream.CloseSend(); err != nil {
+		ss.Log.WithError(err).Warn("failed to close stream send side")
+	}
+
+	ss.Reattestable = resp.GetResult().Reattestable
+	ss.SVID = svid
+	return nil, nil
+}
+
 func findKeyForSVID(keys []keymanager.Key, svid []*x509.Certificate) (keymanager.Key, bool) {
 	if len(svid) == 0 {
 		return nil, false
@@ -277,4 +384,21 @@ func findKeyForSVID(keys []keymanager.Key, svid []*x509.Certificate) (keymanager
 		}
 	}
 	return nil, false
+}
+
+func getSVIDFromAttestAgentResponse(r *agentv1.AttestAgentResponse) ([]*x509.Certificate, error) {
+	if r.GetResult().Svid == nil {
+		return nil, errors.New("attest response is missing SVID")
+	}
+
+	svid, err := x509util.RawCertsToCertificates(r.GetResult().Svid.CertChain)
+	if err != nil {
+		return nil, fmt.Errorf("invalid SVID cert chain: %w", err)
+	}
+
+	if len(svid) == 0 {
+		return nil, errors.New("empty SVID cert chain")
+	}
+
+	return svid, nil
 }
